+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Deprecated
 ### Removed
+
 ### Fixed
+
+- Fixed a bug in trace reporting by `quint run` (#913)
+
 ### Security
 
 ## v0.11.0 -- 2023-05-24